+++ tests/testthat/test-mismatch.R
@@ -11,8 +11,3 @@ test_that("failures are pushed to log file", {
   n_failures <- length(keep(mock_cases, inherits, "mismatch_case"))
   expect_identical(n_logged, n_failures)
 })
-
-test_that("Verbose outputs svg files", {
-  output <- mock_cases_outputs$output
-  expect_true(grepl("Failed doppelganger: myplot", output))
-})